requirements.txt
pyserial
numpy
python-socketio[client]